Improve websocket timeout handling
While websocket.close() is being properly called when pings timeout,
it appears that the close() isn't (always?) interrupting webslocket.recv()

causing relay_messages_to_peer() to not terminate (leaving the
associated WSClient in the meetup_table.
To resolve, added WSClient.communicate_with_peer() that will wait for
either relay_messages_to_peer() to terminate and/or a new
ping_timeout_future - which is satisfied when pings timeout. If the
relay_messages_to_peer() doesn't terminate properly (due to the
recv() issue), the ping_timeout_future will ensure that
communicate_with_peer() will complete and remove the WSClient
from the meetup_table.

Performed a variety of connect, disconnect, and timeout scenarios
to ensure the cases are handled (including a reproductive test case
that demonstrated the issue).